Create a seperate class for book creator
